requirements.txt
streamlit
openai
python-dotenv
psycopg2-binary
passlib
bcrypt
supabase
pytz